// frontend/src/app/api-gen/models/session-detail-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SessionItemInfo } from '../models/session-item-info';
export interface SessionDetailResponse {
  created_at: string;
  env_a: string;
  env_b: string;
  id: string;
  items: Array<SessionItemInfo>;
  service: string;
  status_counts: {
[key: string]: number;
};
  title: string;
}