// frontend/src/app/api-gen/models/session-item-info.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface SessionItemInfo {
  applied_at?: (string | null);
  diff_err?: (string | null);
  diff_json?: (string | null);
  is_secret: boolean;
  name: string;
  notes?: (string | null);
  position: number;
  preview?: (string | null);
  script?: (string | null);
  service: string;
  status: string;
  updated_at?: (string | null);
  visited_at?: (string | null);
}